requirements.txt
celery==5.4.0
faiss_cpu==1.8.0
Flask==3.0.3
Flask_Cors==4.0.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
GitPython==3.1.43
GitPython==3.1.43
langchain==0.1.19
loguru==0.7.2
python-dotenv==1.0.1
pyunpack==0.3
Requests==2.31.0
setuptools==68.2.2
transformers==4.39.1
voyageai==0.2.2
Werkzeug==3.0.3
wheel==0.43.0
